# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anagram Finder API",
        "description": "API for finding anagrams and managing wordbase",
        "contact": {
            "email": "admin@anagramfinder.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/anagrams/{word}": {
            "get": {
                "tags": [
                    "Anagrams"
                ],
                "summary": "Find anagrams for a given word",
                "description": "Returns all anagrams found for the provided word along with metadata",
                "operationId": "findAnagrams",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "description": "The word to find anagrams for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 100,
                            "minLength": 1,
                            "example": "listen"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "word": {
                                                    "type": "string",
                                                    "example": "listen"
                                                },
                                                "anagrams": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "silent",
                                                        "enlist"
                                                    ]
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "search_time_ms": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 12.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Word parameter cannot be empty"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "INVALID_WORD"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Wordbase not available. Please import wordbase first."
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "WORDBASE_NOT_READY"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}